table look improvements
