+++ client_app/.prettierrc
@@ -4,5 +4,5 @@
     "svelteAllowShorthand": false,
     "svelteIndentScriptAndStyle": false,
     "printWidth": 120,
-    "tabWidth": 4
+    "tabWidth": 2
 }
